Update vendored copy of wheel
