+++ setuptools/_vendor/wheel-0.45.1.dist-info/LICENSE.txt
