+++ setuptools/_vendor/wheel/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import annotations
 
-__version__ = "0.43.0"
+__version__ = "0.45.1"
