+++ setuptools/_vendor/wheel/util.py
@@ -5,15 +5,6 @@ import logging
 
 log = logging.getLogger("wheel")
 
-# ensure Python logging is configured
-try:
-    __import__("setuptools.logging")
-except ImportError:
-    # setuptools < ??
-    from . import _setuptools_logging
-
-    _setuptools_logging.configure()
-
 
 def urlsafe_b64encode(data: bytes) -> bytes:
     """urlsafe_b64encode without padding"""
